\name{OptionData-class}
\docType{class}
\alias{OptionData-class}
\alias{initialize,OptionData-method}

\title{\code{OptionData}: A class for options available in the markets.}

\description{
Dataset contains all information about the option derivatives available on the markets.
}


\section{Objects from the Class}{
This object is built by the function \code{DataOpt}.
}
\section{Slots}{
  \describe{
    \item{\code{UnderPrice}:}{ }
    \item{\code{PriceOpt}:}{}
    \item{\code{ImpliedVol}:}{ }
    \item{\code{TimeToMat}:}{ } 
    \item{\code{Strike}:}{}
    \item{\code{Type}:}{}
    \item{\code{rate}:}{ }
    \item{\code{qyield}:}{}
    \item{\code{dateobs}:}{}
      }
}



