\name{PricingMixedTS-package}
\alias{PricingMixedTS-package}
\alias{PricingMixedTS}
\docType{package}
\title{Option Pricing with exponential MixedTS process
}
\description{
The package provides several routines for option pricing with general Levy process. Specific methods are available when the underlying asset prices is an exponential MixedTS process.
}
\details{
\tabular{ll}{
Package: \tab PricingMixedTS\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-03-17\cr
License: \tab What license is it under?\cr
}
}
